requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

cliff!=1.16.0,!=1.17.0,>=1.15.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
pbr>=1.6 # Apache-2.0
GitPython>=1.0.1 # BSD License (3 clause)
PyYAML>=3.1.0 # MIT
six>=1.9.0 # MIT
jsonschema>=2.0.0,<3.0.0,!=2.5.0 # MIT
